// src/models/summary_format.ts
export interface SummaryFormat {
  count_num: number;
  count_distinct: number;
  source_num: number;
  source_distinct: number;
  destination_num: number;
  destination_distinct: number;
  time_num: number;
  time_distinct: number;
  protocol_num: number;
  protocol_distinct: number;
  length_num: number;
  length_distinct: number;
}